test: migrate ClassicAssert to Assert.That, merge *ExtraTests, fix fixture naming
Mechanical cleanups for Beutl.UnitTests Phase 2:

- Migrate all 97 ClassicAssert.* call sites across 10 fixtures to the
  constraint-model Assert.That API and drop the now-unused
  `using NUnit.Framework.Legacy;` import from each. ClassicAssert.Throws<T>
  becomes the direct Assert.Throws<T> form.
- Merge the 7 *ExtraTests fixtures into their primary fixtures (moving all
  [Test] methods and private nested helpers, reconciling usings), then delete
  the Extra files: JsonConverters, ColorMatrix, Color, TimeRange, PixelRect,
  RefStringTokenizer, and RefUtf8StringTokenizer (Utilities primary).
- Fix fixture naming drift: rename TransformParser.cs -> TransformParserTests.cs,
  and resolve the duplicate RefUtf8StringTokenizerTests class name by renaming
  the Engine ParseCommaSeparated/ParseDirty fixture (file and class) to
  RefUtf8StringTokenizerParseTests; the Utilities Read* fixture keeps its name.

Behavior-preserving: all migrated tests still compile and pass.
